requirements.txt
python-dotenv==0.21.0
google-genai==1.2.0
groq==0.18.0
yfinance==0.2.52
phidata==2.7.10